Change coordinate lat lng to decimal and update seed data

+++ Models/Coordinate.cs
@@ -4,7 +4,7 @@ namespace azimuth_api.Models {
     public class Coordinate {
         [Key]
         public long Id {get;set;}
-        public float Lat {get;set;}
-        public float Lng {get;set;}
+        public decimal Lat {get;set;}
+        public decimal Lng {get;set;}
     }
 }
